[0.6.0.220]

+++ src/extends/res/shared/varnishd.py
@@ -54,12 +54,14 @@ class varnishd_shared(interface.resource,res_utls):
                 EXTRAS     = self.extras
                 )
         shexec.execmd(cmd)
+        time.sleep(2)
 
     def _stop(self,context):
-        if impl.rg_utls.get_key("Are you sure stop Varnishd? (y/N)", context)  == "y" :
-            cmdtpl = "cat $PID | xargs kill ; rm $PID  "
-            cmd = Template(cmdtpl).substitute( PID = self.pid)
-            shexec.execmd(cmd)
+        if os.path.exists(self.pid) :
+            if impl.rg_utls.get_key("Are you sure stop Varnishd? (y/N)", context)  == "y" :
+                cmdtpl = "cat $PID | xargs kill ; rm $PID  "
+                cmd = Template(cmdtpl).substitute( PID = self.pid)
+                shexec.execmd(cmd)
 
     def _reload(self,context):
         confname = "vcl_%s_%d" %(time.strftime("%H_%M_%S",time.localtime()) , int(random.random() * 100 ) )
